modified:   .gitignore 	modified:   Makefile 	modified:   iso/boot/grub/grub.cfg 	modified:   src/boot/boot.asm 	modified:   src/kernel/cpu/gdt.c 	modified:   src/kernel/drivers/ata.c 	modified:   src/kernel/drivers/io_port.c 	new file:   src/kernel/fs/devfs.c 	new file:   src/kernel/fs/devices/null.c 	new file:   src/kernel/fs/devices/zero.c 	new file:   src/kernel/fs/msdospart.c 	new file:   src/kernel/fs/vfs.c 	modified:   src/kernel/include/cpu/gdt.h 	modified:   src/kernel/include/display.h 	modified:   src/kernel/include/drivers/ata.h 	modified:   src/kernel/include/drivers/io_port.h 	modified:   src/kernel/include/drivers/keyboard.h 	modified:   src/kernel/include/drivers/rtc.h 	new file:   src/kernel/include/fs/fs.h 	new file:   src/kernel/include/fs/msdospart.h 	modified:   src/kernel/kernel.c 	modified:   src/kernel/memory/mmu.c 	modified:   src/kernel/memory/paging.c 	new file:   src/libc/include/hashmap.h 	new file:   src/libc/include/list.h 	modified:   src/libc/include/string.h 	new file:   src/libc/include/tree.h 	new file:   src/libc/src/dsa/hashmap.c 	new file:   src/libc/src/dsa/list.c 	new file:   src/libc/src/dsa/tree.c 	modified:   src/libc/src/string.c

+++ src/kernel/include/drivers/rtc.h
@@ -28,7 +28,7 @@ typedef enum {
 } cmos_reg;
 
 typedef struct{
-	uint8_t year, mnth, day, hour, min, sec;
+	uint16_t year, mnth, day, hour, min, sec;
 } time_t;
 
 /* function declarations */
